Add image_tag to Attendee#index view
I noticed that the image_tag has been added to Attendee#show, but not
yet to Attendee#index, causing the application to display just the
value of the picture attribute (e.g. the path where the uploaded file
is stored) instead of the image itself in the Attendee#index view.

As the page with the map will look much fancier with the users pictures,
I added image_tag there, too.